Update objects.py
When we submitted a file generated with this lib to swedbank. 
It failed for a format reason. After some research, it appears a CR/LF issue.
The proposed PR fixed our issue. If this is not causing side effect for other bank institution, I will propose to add this change.